+++ src/main/java/io/wiretap/http/message/settings/body/BodyParserConfiguration.java
@@ -5,6 +5,7 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+import java.util.List;
 import java.util.Optional;
 
 @Configuration
@@ -12,7 +13,10 @@ public class BodyParserConfiguration {
 
     @ConditionalOnMissingBean
     @Bean
-    public BodyParser bodyParser(@Autowired Optional<HttpBodyMaskingHandler> maskingHandler) {
-        return new DefaultBodyParser(maskingHandler.orElse(null));
+    public BodyParser bodyParser(
+            @Autowired Optional<HttpBodyMaskingHandler> maskingHandler,
+            @Autowired List<HttpBodyMasker> bodyMaskers
+    ) {
+        return new DefaultBodyParser(maskingHandler.orElse(null), bodyMaskers);
     }
 }